DB updates now feature progress bars. I'm not sure the text color is right on froyo (looked off in the emulator, anyway), but it otherwise works.

+++ trunk/MTGFamiliar/src/com/gelakinetic/mtgfam/activities/MainActivity.java
@@ -322,7 +322,7 @@ public class MainActivity extends FamiliarActivity {
 			case R.id.checkUpdate:
 				// Set the last legality update time back to zero on a forced update
 				SharedPreferences.Editor editor = preferences.edit();
-				editor.putLong("lastLegalityUpdate", 0);
+				editor.putInt("lastLegalityUpdate", 0);
 				editor.commit();
 
                 startService(new Intent(this, DbUpdaterService.class));
